cleanup after rebase
