+++ Makefile
@@ -29,6 +29,10 @@ clean:
 	rm -rf patch-db/client/node_modules
 	rm -rf patch-db/client/dist
 
+	rm backed/workspaces/js_engine/src/artifacts/ARM_JS_SNAPSHOT.bin
+	rm backed/workspaces/js_engine/src/artifacts/JS_SNAPSHOT.bin
+	touch backend/workspaces/snapshot-creator/Cargo.toml
+
 eos.img: $(EMBASSY_SRC) system-images/compat/compat.tar system-images/utils/utils.tar $(EMBASSY_V8_SNAPSHOTS)
 	! test -f eos.img || rm eos.img
 	if [ "$(NO_KEY)" = "1" ]; then NO_KEY=1 ./build/make-image.sh; else ./build/make-image.sh; fi
@@ -51,10 +55,11 @@ product_key.txt:
 	if [ "$(KEY)" != "" ]; then $(shell which echo) -n "$(KEY)" > product_key.txt; fi
 	echo >> product_key.txt
 
-$(EMBASSY_V8_SNAPSHOTS): $(BACKEND_SRC) $(EMBASSY_BINS)
-	cd backend && ./build-arm-v8-snapshot.sh
+$(EMBASSY_V8_SNAPSHOTS): backend/workspaces/snapshot-creator/Cargo.toml
+	cd backend/workspaces/  && ./build-arm-v8-snapshot.sh
+	cd backend/workspaces/  && ./build-v8-snapshot.sh
 
-$(EMBASSY_BINS): $(BACKEND_SRC)
+$(EMBASSY_BINS): $(BACKEND_SRC) $(EMBASSY_V8_SNAPSHOTS)
 	cd backend && ./build-prod.sh
 
 frontend/node_modules: frontend/package.json
