+++ backend/Cargo.toml
@@ -25,10 +25,6 @@ name = "embassyd"
 path = "src/bin/embassyd.rs"
 
 [[bin]]
-name = "create-js-snapshots"
-path = "src/bin/build-js-snapshots.rs"
-
-[[bin]]
 name = "embassy-init"
 path = "src/bin/embassy-init.rs"
 
@@ -65,8 +61,6 @@ color-eyre = "0.5"
 cookie_store = "0.15.0"
 digest = "0.9.0"
 divrem = "1.0.0"
-deno_core = "0.135.0"
-deno_ast = { version = "0.14.1", features = ["transpiling"] }
 ed25519-dalek = { version = "1.0.1", features = ["serde"] }
 emver = { version = "0.1.6", features = ["serde"] }
 fd-lock-rs = "0.1.3"
@@ -87,8 +81,11 @@ libc = "0.2.103"
 log = "0.4.14"
 nix = "0.23.0"
 nom = "7.0.0"
+helpers = {path = "./workspaces/helpers"}
 num = "0.4.0"
 num_enum = "0.5.4"
+models = {version = "*", path = "./workspaces/models"}
+js_engine = {path = './workspaces/js_engine'}
 openssh-keys = "0.5.0"
 openssl = { version = "0.10.36", features = ["vendored"] }
 patch-db = { version = "*", path = "../patch-db/patch-db", features = [
