+++ backend/src/version/mod.rs
@@ -11,6 +11,7 @@ mod v0_3_0;
 mod v0_3_0_1;
 mod v0_3_0_2;
 mod v0_3_0_3;
+mod v0_3_1;
 
 pub type Current = v0_3_0_3::Version;
 
@@ -21,6 +22,7 @@ enum Version {
     V0_3_0_1(Wrapper<v0_3_0_1::Version>),
     V0_3_0_2(Wrapper<v0_3_0_2::Version>),
     V0_3_0_3(Wrapper<v0_3_0_3::Version>),
+    V0_3_1(Wrapper<v0_3_1::Version>),
     Other(emver::Version),
 }
 
@@ -40,6 +42,7 @@ impl Version {
             Version::V0_3_0_1(Wrapper(x)) => x.semver(),
             Version::V0_3_0_2(Wrapper(x)) => x.semver(),
             Version::V0_3_0_3(Wrapper(x)) => x.semver(),
+            Version::V0_3_1(Wrapper(x)) => x.semver(),
             Version::Other(x) => x.clone(),
         }
     }
@@ -154,6 +157,7 @@ pub async fn init<Db: DbHandle>(
         Version::V0_3_0_1(v) => v.0.migrate_to(&Current::new(), db, receipts).await?,
         Version::V0_3_0_2(v) => v.0.migrate_to(&Current::new(), db, receipts).await?,
         Version::V0_3_0_3(v) => v.0.migrate_to(&Current::new(), db, receipts).await?,
+        Version::V0_3_1(v) => v.0.migrate_to(&Current::new(), db, receipts).await?,
         Version::Other(_) => {
             return Err(Error::new(
                 eyre!("Cannot downgrade"),
@@ -189,6 +193,7 @@ mod tests {
             Just(Version::V0_3_0_1(Wrapper(v0_3_0_1::Version::new()))),
             Just(Version::V0_3_0_2(Wrapper(v0_3_0_2::Version::new()))),
             Just(Version::V0_3_0_3(Wrapper(v0_3_0_3::Version::new()))),
+            Just(Version::V0_3_1(Wrapper(v0_3_1::Version::new()))),
             em_version().prop_map(Version::Other),
         ]
     }
