+++ backend/workspaces/Cargo.toml
@@ -1,0 +1,8 @@
+[workspace]
+
+members = [
+    "snapshot-creator",
+    "models",
+    "js_engine",
+    "helpers"
+]
