+++ backend/workspaces/helpers/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "helpers"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+pin-project = "1.0.8"
+tokio = { version = "1.15.*", features = ["full"] }
