+++ backend/workspaces/js_engine/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "js_engine"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+dashmap = "5.1.0"
+deno_core =  "0.136.0"
+deno_ast = {version="0.15.0", features = ["transpiling"]}
+models = {path = "../models"}
+helpers = {path = "../helpers"}
+serde = { version = "1.0.*", features = ["derive", "rc"] }
+serde_json = "1.0.*"
+tokio = { version = "1.*.*", features = ["full"] }
+tracing = "0.1"
