+++ backend/workspaces/models/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "models"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+patch-db = { version = "*", path = "../../../patch-db/patch-db", features = [
+  "trace",
+] }
+serde = { version = "1.0.*", features = ["derive", "rc"] }
+thiserror = "1.0.*"
+emver = { version = "0.1.6", features = ["serde"] }
+rand = "0.7.*"
