+++ backend/workspaces/models/src/invalid_id.rs
@@ -1,0 +1,4 @@
+
+#[derive(Debug, thiserror::Error)]
+#[error("Invalid ID")]
+pub struct InvalidId;
