+++ backend/workspaces/snapshot-creator/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "snapshot-creator"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+dashmap = "5.1.0"
+deno_core =  "0.136.0"
+deno_ast = {version="0.15.0", features = ["transpiling"]}
