+++ system-images/compat/Cargo.toml
@@ -10,6 +10,7 @@ version = "0.1.0"
 anyhow = { version = "1.0.40", features = ["backtrace"] }
 beau_collector = "0.2.1"
 clap = "2.33.3"
+dashmap = "5.3.2"
 embassy-os = { path = "../../backend", default-features = false }
 emver = { version = "0.1.2", features = ["serde"] }
 failure = "0.1.8"
